+++ estime-api/src/main/java/com/estime/common/sse/application/SseSubscriptionManager.java
@@ -1,7 +1,6 @@
 package com.estime.common.sse.application;
 
 import com.estime.common.sse.domain.SseConnection;
-import com.estime.common.sse.domain.SseConnectionManager;
 import com.estime.room.domain.vo.RoomSession;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
